Revert "Merge pull request #482 from maskaravivek/refactorNearby"
This reverts commit 397f7dbc85a0761485fbcb8fbdbd6e595f2c54ca, reversing
changes made to ce994c4f3e684b0c8391fe2ad02d9b67769a24e0.

Revert merge of Nearby refactor.
