--- app/src/main/java/fr/free/nrw/commons/ViewHolder.java
@@ -1,7 +1,0 @@
-package fr.free.nrw.commons;
-
-import android.content.Context;
-
-public interface ViewHolder<T> {
-    void bindModel(Context context, T model);
-}
